mps_comm: only dispatch instruction when a new one arrives
The node vars of the instruction are not always set in order. Sometimes
the error does not arrive at all and others the order they arrive is
missed. To remedy that, send the dispatch the instruction after a new
action arrives. And allways reset the instruction to hearbeat after
dispatching